Unified test migration for table.insertIgnoreMany()
