--- test/run-time/input/table/insert-ignore-many/zero.ts
@@ -1,61 +1,0 @@
-import * as tm from "type-mapping";
-import * as tape from "tape";
-import * as tsql from "../../../../../dist";
-import {Pool} from "../../sql-web-worker/promise.sql";
-import {SqliteWorker} from "../../sql-web-worker/worker.sql";
-
-tape(__filename, async (t) => {
-    const pool = new Pool(new SqliteWorker());
-
-    const test = tsql.table("test")
-        .addColumns({
-            testId : tm.mysql.bigIntUnsigned(),
-            testVal : tm.mysql.bigIntUnsigned(),
-        })
-        .setPrimaryKey(columns => [columns.testId]);
-
-    const insertResult = await pool.acquire(async (connection) => {
-        await connection.exec(`
-            CREATE TABLE test (
-                testId INT PRIMARY KEY,
-                testVal INT
-            );
-        `);
-
-        return test.insertIgnoreMany(
-            connection,
-            []
-        );
-    });
-    t.deepEqual(
-        insertResult.insertedRowCount,
-        BigInt(0)
-    );
-    t.deepEqual(
-        insertResult.warningCount,
-        BigInt(0)
-    );
-    t.deepEqual(
-        insertResult.message,
-        "No rows to insert"
-    );
-
-    await pool
-        .acquire(async (connection) => {
-            return tsql.from(test)
-                .select(columns => [columns])
-                .orderBy(columns => [
-                    columns.testId.asc(),
-                ])
-                .fetchAll(connection);
-        })
-        .then((rows) => {
-            t.deepEqual(
-                rows,
-                []
-            );
-        });
-
-    await pool.disconnect();
-    t.end();
-});
