+++ unified-test/input/_from-ununified/table/insert-ignore-many/insert-disabled.ts
@@ -1,0 +1,74 @@
+import * as tm from "type-mapping";
+import {Test} from "../../../../test";
+import * as tsql from "../../../../../dist";
+
+export const test : Test = ({tape, pool, createTemporarySchema}) => {
+    tape(__filename, async (t) => {
+        const test = tsql.table("test")
+            .addColumns({
+                testId : tm.mysql.bigIntUnsigned(),
+                testVal : tm.mysql.bigIntUnsigned(),
+            })
+            .setPrimaryKey(columns => [columns.testId])
+            .disableInsert();
+
+        await pool.acquire(async (connection) => {
+            await createTemporarySchema(
+                connection,
+                {
+                    tables : [
+                        {
+                            tableAlias : "test",
+                            columns : [
+                                {
+                                    columnAlias : "testId",
+                                    dataType : {
+                                        typeHint : tsql.TypeHint.BIGINT_SIGNED,
+                                    },
+                                },
+                                {
+                                    columnAlias : "testVal",
+                                    dataType : {
+                                        typeHint : tsql.TypeHint.BIGINT_SIGNED,
+                                    },
+                                },
+                            ],
+                            primaryKey : {
+                                multiColumn : false,
+                                columnAlias : "testId",
+                                autoIncrement : false,
+                            },
+                        }
+                    ]
+                }
+            );
+
+            return (test as (typeof test & tsql.InsertableTable)).insertIgnoreMany(
+                connection,
+                [
+                    {
+                        testId : BigInt(4),
+                        testVal : BigInt(400),
+                    }
+                ]
+            ).then(() => {
+                t.fail("Should not be able to insert");
+            }).catch(() => {
+                t.pass("Should fail to insert");
+            });
+        });
+
+        await pool
+            .acquire(async (connection) => {
+                return tsql.from(test).select(columns => [columns]).fetchAll(connection);
+            })
+            .then((rows) => {
+                t.deepEqual(
+                    rows,
+                    []
+                );
+            });
+
+        t.end();
+    });
+};
