+++ unified-test/input/_from-ununified/table/insert-ignore-many/partial-ignore-multi-constraint.ts
@@ -1,0 +1,120 @@
+import * as tm from "type-mapping";
+import {Test} from "../../../../test";
+import * as tsql from "../../../../../dist";
+
+export const test : Test = ({tape, pool, createTemporarySchema}) => {
+    tape(__filename, async (t) => {
+        const test = tsql.table("test")
+            .addColumns({
+                testId : tm.mysql.bigIntUnsigned(),
+                testVal : tm.mysql.bigIntUnsigned(),
+            })
+            .setPrimaryKey(columns => [columns.testId]);
+
+        const insertResult = await pool.acquire(async (connection) => {
+            await createTemporarySchema(
+                connection,
+                {
+                    tables : [
+                        {
+                            tableAlias : "test",
+                            columns : [
+                                {
+                                    columnAlias : "testId",
+                                    dataType : {
+                                        typeHint : tsql.TypeHint.BIGINT_SIGNED,
+                                    },
+                                },
+                                {
+                                    columnAlias : "testVal",
+                                    dataType : {
+                                        typeHint : tsql.TypeHint.BIGINT_SIGNED,
+                                    },
+                                },
+                            ],
+                            primaryKey : {
+                                multiColumn : false,
+                                columnAlias : "testId",
+                                autoIncrement : false,
+                            },
+                            candidateKeys : [
+                                ["testVal"],
+                            ],
+                        }
+                    ]
+                }
+            );
+
+            await test.insertMany(
+                connection,
+                [
+                    {
+                        testId : BigInt(0),
+                        testVal : BigInt(100),
+                    },
+                    {
+                        testId : BigInt(2),
+                        testVal : BigInt(222),
+                    },
+                ]
+            );
+
+            return test.insertIgnoreMany(
+                connection,
+                [
+                    {
+                        testId : BigInt(1),
+                        testVal : BigInt(100),
+                    },
+                    {
+                        testId : BigInt(2),
+                        testVal : BigInt(200),
+                    },
+                    {
+                        testId : BigInt(3),
+                        testVal : BigInt(300),
+                    },
+                ]
+            );
+        });
+        t.deepEqual(
+            insertResult.insertedRowCount,
+            BigInt(1)
+        );
+        t.deepEqual(
+            insertResult.warningCount,
+            BigInt(2)
+        );
+
+        await pool
+            .acquire(async (connection) => {
+                return tsql.from(test)
+                    .select(columns => [columns])
+                    .orderBy(columns => [
+                        columns.testId.asc(),
+                    ])
+                    .fetchAll(connection);
+            })
+            .then((rows) => {
+                t.deepEqual(
+                    rows,
+                    [
+                        {
+                            testId : BigInt(0),
+                            testVal : BigInt(100),
+                        },
+                        {
+                            testId : BigInt(2),
+                            testVal : BigInt(222),
+                        },
+                        {
+                            testId : BigInt(3),
+                            testVal : BigInt(300),
+                        },
+                    ]
+                );
+            });
+
+        t.end();
+    });
+};
