+++ unified-test/input/_from-ununified/table/insert-ignore-many/zero.ts
@@ -1,0 +1,81 @@
+import * as tm from "type-mapping";
+import {Test} from "../../../../test";
+import * as tsql from "../../../../../dist";
+
+export const test : Test = ({tape, pool, createTemporarySchema}) => {
+    tape(__filename, async (t) => {
+        const test = tsql.table("test")
+            .addColumns({
+                testId : tm.mysql.bigIntUnsigned(),
+                testVal : tm.mysql.bigIntUnsigned(),
+            })
+            .setPrimaryKey(columns => [columns.testId]);
+
+        const insertResult = await pool.acquire(async (connection) => {
+            await createTemporarySchema(
+                connection,
+                {
+                    tables : [
+                        {
+                            tableAlias : "test",
+                            columns : [
+                                {
+                                    columnAlias : "testId",
+                                    dataType : {
+                                        typeHint : tsql.TypeHint.BIGINT_SIGNED,
+                                    },
+                                },
+                                {
+                                    columnAlias : "testVal",
+                                    dataType : {
+                                        typeHint : tsql.TypeHint.BIGINT_SIGNED,
+                                    },
+                                },
+                            ],
+                            primaryKey : {
+                                multiColumn : false,
+                                columnAlias : "testId",
+                                autoIncrement : false,
+                            },
+                        }
+                    ]
+                }
+            );
+
+            return test.insertIgnoreMany(
+                connection,
+                []
+            );
+        });
+        t.deepEqual(
+            insertResult.insertedRowCount,
+            BigInt(0)
+        );
+        t.deepEqual(
+            insertResult.warningCount,
+            BigInt(0)
+        );
+        t.deepEqual(
+            insertResult.message,
+            "No rows to insert"
+        );
+
+        await pool
+            .acquire(async (connection) => {
+                return tsql.from(test)
+                    .select(columns => [columns])
+                    .orderBy(columns => [
+                        columns.testId.asc(),
+                    ])
+                    .fetchAll(connection);
+            })
+            .then((rows) => {
+                t.deepEqual(
+                    rows,
+                    []
+                );
+            });
+
+        t.end();
+    });
+};
